--- TicketSelling.Repositories.Tests/TestDataExtensions.cs
@@ -1,16 +1,0 @@
-﻿using TicketSelling.Context.Contracts.Models;
-
-namespace TicketSelling.Repositories.Tests
-{
-    internal static class TestDataExtensions
-    {
-        public static void BaseAuditSetParamtrs<TEntity>(this TEntity entity) where TEntity : BaseAuditEntity
-        {
-            entity.Id = Guid.NewGuid();
-            entity.CreatedAt = DateTimeOffset.UtcNow;
-            entity.CreatedBy = $"CreatedBy{Guid.NewGuid():N}";
-            entity.UpdatedAt = DateTimeOffset.UtcNow;
-            entity.UpdatedBy = $"UpdatedBy{Guid.NewGuid():N}";
-        }
-    }
-}
