+++ TicketSelling.Services.Tests/TestsServices/HallServiceTests.cs
@@ -2,6 +2,7 @@
 using FluentAssertions;
 using TicketSelling.Context.Contracts.Models;
 using TicketSelling.Context.Tests;
+using TicketSelling.Repositories.Contracts.ReadInterfaces;
 using TicketSelling.Repositories.ReadRepositories;
 using TicketSelling.Repositories.WriteRepositoriеs;
 using TicketSelling.Services.AutoMappers;
@@ -14,30 +15,29 @@ using Xunit;
 
 namespace TicketSelling.Services.Tests.Tests
 {
-    public class ClientServiceTest : TicketSellingContextInMemory
+    public class HallServiceTests : TicketSellingContextInMemory
     {
-        private readonly IClientService clientService;
-        private readonly ClientReadRepository clientReadRepository;
+        private readonly IHallService hallService;
+        private readonly HallReadRepository hallReadRepository;
 
         /// <summary>
-        /// Инициализирует новый экземпляр <see cref="ClientServiceTest"/>
+        /// Инициализирует новый экземпляр <see cref="HallServiceTests"/>
         /// </summary>
-        public ClientServiceTest()
+        public HallServiceTests()
         {
             var config = new MapperConfiguration(cfg =>
             {
                 cfg.AddProfile(new ServiceMapper());
             });
 
-            clientReadRepository = new ClientReadRepository(Reader);
-
-            clientService = new ClientService(new ClientWriteRepository(WriterContext), clientReadRepository,
+            hallReadRepository = new HallReadRepository(Reader);
+            hallService = new HallService(new HallWriteRepository(WriterContext), hallReadRepository,
                 UnitOfWork, config.CreateMapper(), new ServicesValidatorService(new CinemaReadRepository(Reader), 
-                clientReadRepository, new FilmReadRepository(Reader), new HallReadRepository(Reader)));
+                new ClientReadRepository(Reader), new FilmReadRepository(Reader), hallReadRepository));
         }
 
         /// <summary>
-        /// Получение <see cref="Client"/> по идентификатору возвращает null
+        /// Получение <see cref="Hall"/> по идентификатору возвращает null
         /// </summary>
         [Fact]
         public async Task GetByIdShouldReturnNull()
@@ -46,26 +46,26 @@ namespace TicketSelling.Services.Tests.Tests
             var id = Guid.NewGuid();
 
             // Act
-            Func<Task> result = () => clientService.GetByIdAsync(id, CancellationToken);
+            Func<Task> result = () => hallService.GetByIdAsync(id, CancellationToken);
 
             // Assert
-            await result.Should().ThrowAsync<TimeTableEntityNotFoundException<Client>>()
-               .WithMessage($"*{id}*");
+            await result.Should().ThrowAsync<TimeTableEntityNotFoundException<Hall>>()
+                .WithMessage($"*{id}*");
         }
 
         /// <summary>
-        /// Получение <see cref="Client"/> по идентификатору возвращает данные
+        /// Получение <see cref="Hall"/> по идентификатору возвращает данные
         /// </summary>
         [Fact]
         public async Task GetByIdShouldReturnValue()
         {
             //Arrange
-            var target = TestDataGenerator.Client();
-            await Context.Clients.AddAsync(target);
+            var target = TestDataGenerator.Hall();
+            await Context.Halls.AddAsync(target);
             await Context.SaveChangesAsync(CancellationToken);
 
             // Act
-            var result = await clientService.GetByIdAsync(target.Id, CancellationToken);
+            var result = await hallService.GetByIdAsync(target.Id, CancellationToken);
 
             // Assert
             result.Should()
@@ -73,22 +73,19 @@ namespace TicketSelling.Services.Tests.Tests
                 .And.BeEquivalentTo(new
                 {
                     target.Id,
-                    target.Age,
-                    target.FirstName,
-                    target.LastName,
-                    target.Email,
-                    target.Patronymic
+                    target.NumberOfSeats,
+                    target.Number
                 });
         }
 
         /// <summary>
-        /// Получение <see cref="IEnumerable{Client}"/> по идентификаторам возвращает пустую коллекцию
+        /// Получение <see cref="IEnumerable{Hall}"/> по идентификаторам возвращает пустую коллекцию
         /// </summary>
         [Fact]
         public async Task GetAllShouldReturnEmpty()
         {
             // Act
-            var result = await clientService.GetAllAsync(CancellationToken);
+            var result = await hallService.GetAllAsync(CancellationToken);
 
             // Assert
             result.Should()
@@ -97,20 +94,20 @@ namespace TicketSelling.Services.Tests.Tests
         }
 
         /// <summary>
-        /// Получение <see cref="IEnumerable{Client}"/> по идентификаторам возвращает данные
+        /// Получение <see cref="IEnumerable{Hall}"/> по идентификаторам возвращает данные
         /// </summary>
         [Fact]
         public async Task GetAllShouldReturnValues()
         {
             //Arrange
-            var target = TestDataGenerator.Client();
+            var target = TestDataGenerator.Hall();
 
-            await Context.Clients.AddRangeAsync(target,
-                TestDataGenerator.Client(x => x.DeletedAt = DateTimeOffset.UtcNow));
+            await Context.Halls.AddRangeAsync(target,
+                TestDataGenerator.Hall(x => x.DeletedAt = DateTimeOffset.UtcNow));
             await Context.SaveChangesAsync(CancellationToken);
 
             // Act
-            var result = await clientService.GetAllAsync(CancellationToken);
+            var result = await hallService.GetAllAsync(CancellationToken);
 
             // Assert
             result.Should()
@@ -120,7 +117,7 @@ namespace TicketSelling.Services.Tests.Tests
         }
 
         /// <summary>
-        /// Удаление несуществуюущего <see cref="Client"/>
+        /// Удаление несуществуюущего <see cref="Hall"/>
         /// </summary>
         [Fact]
         public async Task DeletingNonExistentCinemaReturnExсeption()
@@ -129,148 +126,146 @@ namespace TicketSelling.Services.Tests.Tests
             var id = Guid.NewGuid();
 
             // Act
-            Func<Task> result = () => clientService.DeleteAsync(id, CancellationToken);
+            Func<Task> result = () => hallService.DeleteAsync(id, CancellationToken);
 
             // Assert
-            await result.Should().ThrowAsync<TimeTableEntityNotFoundException<Client>>()
-               .WithMessage($"*{id}*");
+            await result.Should().ThrowAsync<TimeTableEntityNotFoundException<Hall>>()
+                .WithMessage($"*{id}*");
         }
 
         /// <summary>
-        /// Удаление удаленного <see cref="Client"/>
+        /// Удаление удаленного <see cref="Hall"/>
         /// </summary>
         [Fact]
         public async Task DeletingDeletedCinemaReturnExсeption()
         {
             //Arrange
-            var model = TestDataGenerator.Client(x => x.DeletedAt = DateTime.UtcNow);
-            await Context.Clients.AddAsync(model);
+            var model = TestDataGenerator.Hall(x => x.DeletedAt = DateTime.UtcNow);
+            await Context.Halls.AddAsync(model);
             await Context.SaveChangesAsync(CancellationToken);
 
             // Act
-            Func<Task> result = () => clientService.DeleteAsync(model.Id, CancellationToken);
+            Func<Task> result = () => hallService.DeleteAsync(model.Id, CancellationToken);
 
             // Assert
-            await result.Should().ThrowAsync<TimeTableEntityNotFoundException<Client>>()
+            await result.Should().ThrowAsync<TimeTableEntityNotFoundException<Hall>>()
                 .WithMessage($"*{model.Id}*");
-        }      
+        }        
 
         /// <summary>
-        /// Удаление <see cref="Client"/>
+        /// Удаление <see cref="Hall"/>
         /// </summary>
         [Fact]
         public async Task DeleteShouldWork()
         {
             //Arrange
-            var model = TestDataGenerator.Client();
-            await Context.Clients.AddAsync(model);
+            var model = TestDataGenerator.Hall();
+            await Context.Halls.AddAsync(model);
             await UnitOfWork.SaveChangesAsync(CancellationToken);
 
             //Act
-            Func<Task> act = () => clientService.DeleteAsync(model.Id, CancellationToken);
+            Func<Task> act = () => hallService.DeleteAsync(model.Id, CancellationToken);
 
             // Assert
             await act.Should().NotThrowAsync();
-            var entity = Context.Clients.Single(x => x.Id == model.Id);
+            var entity = Context.Halls.Single(x => x.Id == model.Id);
             entity.Should().NotBeNull();
             entity.DeletedAt.Should().NotBeNull();
         }
 
         /// <summary>
-        /// Добавление <see cref="Client"/>
+        /// Добавление <see cref="Hall"/>
         /// </summary>
         [Fact]
         public async Task AddShouldWork()
         {
             //Arrange
-            var model = TestDataGenerator.ClientModel();
+            var model = TestDataGenerator.HallModel();
 
             //Act
-            Func<Task> act = () => clientService.AddAsync(model, CancellationToken);
+            Func<Task> act = () => hallService.AddAsync(model, CancellationToken);
 
             // Assert
             await act.Should().NotThrowAsync();
-            var entity = Context.Clients.Single(x => x.Id == model.Id);
+            var entity = Context.Halls.Single(x => x.Id == model.Id);
             entity.Should().NotBeNull();
             entity.DeletedAt.Should().BeNull();
         }
 
         /// <summary>
-        /// Добавление невалидируемого <see cref="Client"/>
+        /// Добавление не валидируемого <see cref="Hall"/>
         /// </summary>
         [Fact]
         public async Task AddShouldValidationException()
         {
             //Arrange
-            var model = TestDataGenerator.ClientModel(x => x.FirstName = "T");
+            var model = TestDataGenerator.HallModel(x => x.NumberOfSeats = -1);
 
             //Act
-            Func<Task> act = () => clientService.AddAsync(model, CancellationToken);
+            Func<Task> act = () => hallService.AddAsync(model, CancellationToken);
 
             // Assert
             await act.Should().ThrowAsync<TimeTableValidationException>();
         }
 
         /// <summary>
-        /// Изменение несуществующего <see cref="Client"/>
+        /// Изменение несуществующего <see cref="Hall"/>
         /// </summary>
         [Fact]
         public async Task EditShouldNotFoundException()
         {
             //Arrange
-            var model = TestDataGenerator.ClientModel();
+            var model = TestDataGenerator.HallModel();
 
             //Act
-            Func<Task> act = () => clientService.EditAsync(model, CancellationToken);
+            Func<Task> act = () => hallService.EditAsync(model, CancellationToken);
 
             // Assert
-            await act.Should().ThrowAsync<TimeTableEntityNotFoundException<Client>>()
+            await act.Should().ThrowAsync<TimeTableEntityNotFoundException<Hall>>()
                 .WithMessage($"*{model.Id}*");
         }
 
         /// <summary>
-        /// Изменение невалидируемого <see cref="Client"/>
+        /// Изменение невалидируемого <see cref="Hall"/>
         /// </summary>
         [Fact]
         public async Task EditShouldValidationException()
         {
             //Arrange
-            var model = TestDataGenerator.ClientModel(x => x.FirstName = "T");
+            var model = TestDataGenerator.HallModel(x => x.NumberOfSeats = -1);
 
             //Act
-            Func<Task> act = () => clientService.EditAsync(model, CancellationToken);
+            Func<Task> act = () => hallService.EditAsync(model, CancellationToken);
 
             // Assert
             await act.Should().ThrowAsync<TimeTableValidationException>();
         }
 
         /// <summary>
-        /// Изменение <see cref="Client"/>
+        /// Изменение <see cref="Hall"/>
         /// </summary>
         [Fact]
         public async Task EditShouldWork()
         {
             //Arrange
-            var model = TestDataGenerator.ClientModel();
-            var client = TestDataGenerator.Client(x => x.Id = model.Id);
-            await Context.Clients.AddAsync(client);
+            var model = TestDataGenerator.HallModel();
+            var hall = TestDataGenerator.Hall(x => x.Id = model.Id);
+            await Context.Halls.AddAsync(hall);
             await UnitOfWork.SaveChangesAsync(CancellationToken);
 
             //Act
-            Func<Task> act = () => clientService.EditAsync(model, CancellationToken);
+            Func<Task> act = () => hallService.EditAsync(model, CancellationToken);
 
             // Assert
             await act.Should().NotThrowAsync();
-            var entity = Context.Clients.Single(x => x.Id == client.Id);
+            var entity = Context.Halls.Single(x => x.Id == hall.Id);
             entity.Should().NotBeNull()
                 .And
                 .BeEquivalentTo(new
                 {
                     model.Id,
-                    model.FirstName,
-                    model.LastName,
-                    model.Patronymic,
-                    model.Email
+                    model.Number,
+                    model.NumberOfSeats
                 });
         }
     }
